+++ domain/models/calculated_tax.py
@@ -1,9 +1,9 @@
 from dataclasses import dataclass
-
+from decimal import Decimal
 
 @dataclass(frozen=True)
 class CalculatedTax:
     nome: str
-    aliquota: float
-    base_calculo: float
-    valor: float
+    aliquota: Decimal
+    base_calculo: Decimal
+    valor: Decimal
